requirements.txt
pymongo
python-dotenv
snscrape
beautifulsoup4
requests
tika
watchdog
elasticsearch
certifi